requirements.txt
numpy
matplotlib~=3.7.2
pandas~=2.0.3
requests
openpyxl~=3.1.2
xlrd
seaborn
radon

langchain_groq
langchain_core
python_dotenv